Move `PhpWordTemplateProcessor` to namespace `Civi\Civioffice\PhpWord`
